Add NGAP message handling and server configuration
- Introduced server configuration for PWS-IWS in pwsiws.yaml.
- Enhanced app initialization to register NGAP log domain.
- Updated context.h to change serial_number type to uint16_t for better compatibility.
- Added ngap-path.c to meson.build for NGAP message processing.
- Refactored namf-build.c to streamline non-UE N2 message transfer handling.
- Improved sbcap-handler.c to manage warning context and message broadcasting.

+++ src/pwsiws/context.h
@@ -78,7 +78,7 @@ typedef struct pwsiws_warning_s {
     /* Warning message content */
     struct {
         uint8_t     message_type;       /* Warning message type */
-        uint8_t     serial_number;      /* Serial number */
+        uint16_t    serial_number;      /* Serial number */
         uint8_t     message_identifier[50]; /* Message identifier */
         uint8_t     data_coding_scheme; /* Data coding scheme */
         uint8_t     warning_security_information[50]; /* Security info */
